Added test_samples.py, renamed units tests to test_pyrandom2pm, updated __main__.py
