+++ Pipfile
@@ -9,7 +9,7 @@ build = "*"
 twine = "*"
 actions-python-github = "*"
 coverage = "*"
-pyrandom2pm = {file = ".", editable = true}
+pyrandom2pm = "*"
 
 [dev-packages]
